backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "google-generativeai>=0.8.6",
    "pyrate-limiter==3.6.0",
    "python-dotenv>=1.2.1",
    "sec-edgar-downloader==5.0.3",
    "yfinance>=1.0",
]
